deprecated/v1/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
mtgsdk==1.0.0
ollama==0.1.6
PyPDF2==3.0.1 